csv2pcap: support for coded phy and include power level
Add support for Coded Phy in the csv2pcap conversion script.

Unlike the BLE 1M and BLE 2M packets, CODED PHY packets contain
two entries in results/<sim_id>/*.Tx.csv files (one for FEC1 and a
second for FEC2). This modification takes this into account when
creating a BLE entry for Wireshark.
This modification also adds the power of transmissions for Wireshark.

Signed-off-by: Louka Yerly <louka.yerly@hes-so.ch>
Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>